requirements.txt
appnope==0.1.0
astroid==2.0.4
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.0.2
boto3==1.9.64
botocore==1.12.64
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
coloredlogs==10.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
docutils==0.14
en-core-web-md==2.0.0
entrypoints==0.2.3
fakeredis==0.10.3
Flask==1.0.2
Flask-API==1.1
Flask-WTF==0.14.2
future==0.16.0
gevent==1.3.7
greenlet==0.4.15
h5py==2.8.0
humanfriendly==4.17
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
klein==17.10.0
lazy-object-proxy==1.3.1
lxml==4.3.0
MarkupSafe==1.1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==0.28.0
music21==5.5.0
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
notebook==5.7.2
numpy==1.14.5
oauthlib==3.0.0
opencv-python==3.4.3.18
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pip-autoremove==0.9.1
plac==0.9.6
preshed==1.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.1.0
ptyprocess==0.6.0
Pygments==2.3.0
PyHamcrest==1.9.0
pylint==2.1.1
pyparsing==2.3.0
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
rasa-nlu==0.13.8
redis==2.10.6
regex==2017.4.5
requests==2.18.4
requests-oauthlib==1.2.0
ruamel.yaml==0.15.37
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
soupsieve==1.7.3
spacy==2.0.12
tabulate==0.8.2
tensorflow==0.12.0
terminado==0.8.1
testpath==0.4.2
thinc==6.10.3
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
tweepy==3.7.0
Twisted==18.9.0
typing==3.6.4
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.10.11
WTForms==2.2.1
zope.interface==4.6.0
